ECC-2210: Tools: bufr_ls with '-j' and empty output will generate invalid JSON

+++ tests/bufr_ls_json.sh
@@ -41,5 +41,17 @@ do
 done
 
 
+# ECC-2210: bufr_ls with '-j' and empty output
+rm -f $tempLog
+input=aaen_55.bufr
+${tools_dir}/bufr_ls -j -w localTablesVersionNumber=42 $input > $tempLog
+if [ -s $tempLog ]; then
+  # File exists and has a size greater than zero
+  echo "ERROR: JSON output should have been empty!" >&2
+  echo "See temp file $tempLog"
+  exit 1
+fi
+
+
 # Clean up
 rm -f $tempLog
